TEST NAVEGACION CENSO ACTUALIZADOS
Se han movido los tests de navegacion de exportar a un nuevo fichero.py
Se han creado test para la importacion de ficheros en selenium, son tests de navegacion

Parrado Gordon, Raul <@raupargor>
