+++ decide/census/templates/censusImport.html
@@ -25,8 +25,8 @@
   
   </form>
 
-<h3 style=color:red>{{vacio}}</h3>
-<h3 style=color:green>{{importado}}</h3>
+<h3 id=vacioTest style=color:red>{{vacio}}</h3>
+<h3 id=importadoTest style=color:green>{{importado}}</h3>
 
     <table>
     <tr>
