+++ decide/census/testFiles/censo.csv
@@ -1,0 +1,4 @@
+2,2
+2,1
+1,2
+1,1
