+++ decide/census/testFiles/censoIncorrecto.csv
@@ -1,0 +1,5 @@
+e,2
+:,1
+,2
+
+1,1
