+++ decide/census/tests.py
@@ -79,109 +79,6 @@ class CensusTestCase(BaseTestCase):
         self.assertEqual(response.status_code, 204)
         self.assertEqual(0, Census.objects.count())
 
-class ExportCensus(StaticLiveServerTestCase):
-
-    def setUp(self):
-        self.base = BaseTestCase()
-        self.base.setUp()
-        options = webdriver.ChromeOptions()
-        options.headless = True
-        self.driver = webdriver.Chrome(options=options)
-        self.vars = {}
-        self.census=Census(voting_id=1,voter_id=1)
-        self.census.save()
-
-    def tearDown(self):
-        super().tearDown()
-        self.driver.quit()
-
-        self.base.tearDown()
-
-    def test_export_census_positive(self):
-        #Test name: exportar_censo
-        #Step # | name | target | value
-        #1 | open | /admin/ | 
-        self.driver.get(f'{self.live_server_url}/admin/')
-        self.driver.find_element_by_id('id_username').send_keys("adminprueba")
-        self.driver.find_element_by_id('id_password').send_keys("qwerty",Keys.ENTER)
-
-        
-        # 2 | setWindowSize | 821x694 | 
-        self.driver.set_window_size(821, 694)
-
-        # 3 | click | linkText=Censuss | 
-        self.driver.find_element(By.LINK_TEXT, "Censuss").click()
-
-        # 4 | click | id=action-toggle | 
-        self.driver.find_element(By.ID, "action-toggle").click()
-
-        # 5 | select | name=action | label=Export census
-        dropdown = self.driver.find_element(By.NAME, "action")
-        dropdown.find_element(By.XPATH, "//option[. = 'Export census']").click()
-
-        
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).click_and_hold().perform()
-
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).perform()
-
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).release().perform()
-
-        # 6 | click | name=index | 
-        self.driver.find_element(By.NAME, "index").click()
-
-        #Refrescamos la página
-        self.driver.refresh()
-
-        #Si la exportación se ha realizado, debe aparecer un mensaje
-        mensaje = self.driver.find_element(By.CLASS_NAME, "success").text
-
-        self.assertEquals(mensaje, 'Exportación realizada con éxito')
-
-    def test_export_census_negative(self):
-        self.driver.get(f'{self.live_server_url}/admin/')
-        self.driver.find_element_by_id('id_username').send_keys("adminprueba")
-        self.driver.find_element_by_id('id_password').send_keys("qwerty",Keys.ENTER)
-
-        
-        # 2 | setWindowSize | 821x694 | 
-        self.driver.set_window_size(821, 694)
-
-        # 3 | click | linkText=Censuss | 
-        self.driver.find_element(By.LINK_TEXT, "Censuss").click()
-
-        # 4 | click | id=action-toggle | 
-        #self.driver.find_element(By.ID, "action-toggle").click()
-
-        # 5 | select | name=action | label=Export census
-        dropdown = self.driver.find_element(By.NAME, "action")
-        dropdown.find_element(By.XPATH, "//option[. = 'Export census']").click()
-
-        
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).click_and_hold().perform()
-
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).perform()
-
-        element = self.driver.find_element(By.NAME, "action")
-        actions = ActionChains(self.driver)
-        actions.move_to_element(element).release().perform()
-
-        # 6 | click | name=index | 
-        self.driver.find_element(By.NAME, "index").click()
-
-        #Si la exportación se ha realizado, debe aparecer un mensaje
-        mensaje = self.driver.find_element(By.CLASS_NAME, "warning").text
-
-        self.assertEquals(mensaje, 'Items must be selected in order to perform actions on them. No items have been changed.')
 
 class ExportCensusUnitTest(BaseTestCase):
 
@@ -242,10 +139,10 @@ class FilterCensusUnitTest(BaseTestCase):
         user3.save()
 
         #Creación de personas asociadas a censos
-        pers1 = Persona(usuario=user1, sexo='masculino', ip= '127.0.0.1', region='ES', edad=18)
+        pers1 = Persona(usuario=user1, sexo='hombre', ip= '127.0.0.1', region='ES', edad=18)
         pers1.save()
 
-        pers2 = Persona(usuario=user2, sexo='femenino', ip= '127.0.0.1', region='ES', edad=41)
+        pers2 = Persona(usuario=user2, sexo='mujer', ip= '127.0.0.1', region='ES', edad=41)
         pers2.save()
 
         pers3 = Persona(usuario=user3, sexo='otro', ip= '127.0.0.1', region='ES', edad=78)
@@ -274,7 +171,7 @@ class FilterCensusUnitTest(BaseTestCase):
         self.assertEqual(census.voter_id, 3)
 
     def test_sex_filter_negative(self):
-        filter = admin.SexCensusFilter(None, {'genero':'inventado'}, Census, admin.CensusAdmin)
+        filter = admin.SexCensusFilter(None, {'sexo':'inventado'}, Census, admin.CensusAdmin)
         census = filter.queryset(None, Census.objects.all())
         self.assertEqual(census, None)
 
@@ -300,4 +197,33 @@ class FilterCensusUnitTest(BaseTestCase):
     
 
 
-    
+    class ImportCensusUnitTest(BaseTestCase):
+
+        def setUp(self):
+            self.census = Census(voting_id = 1, voter_id=1)
+            self.census.save()
+            user_admin = User(username='admincensus', is_staff=True, is_superuser=True)
+            user_admin.set_password=('qwerty')
+            user_admin.save()
+            self.user_admin = user_admin
+
+            user_noadmin = User(username='simpleuser')
+            user_noadmin.set_password('qwery')
+            user_noadmin.save()
+            self.user_noadmin = user_noadmin
+            super().setUp()
+
+        #Probamos que un usuario administrador exporta un censo y se lo descarga
+        def test_export_census_positive(self):
+            c = Client()
+            c.force_login(self.user_admin)
+            response = c.post("/admin/census/census/", {'action':'export_census', '_selected_action': str(self.census.id)}, format='json')
+            self.assertEqual(response.status_code, 200)
+            self.assertEquals(response.get('Content-Disposition'), 'attachment; filename="censo.csv"')
+
+        #Probamos que un usuario NO administrador intenta acceder al panel de administración y es redirigido al login
+        def test_export_census_negative(self): 
+            c = Client()
+            c.force_login(self.user_noadmin)
+            response = c.post("/admin/census/census/", {'action':'export_census', '_selected_action': str(self.census.id)}, format='json')
+            self.assertEqual(response.status_code, 302)
